Zmiana typu zmiennej ID

+++ src/main/java/com/kodilla/ecommercee/repository/GroupProductRepository.java
@@ -3,5 +3,5 @@ package com.kodilla.ecommercee.repository;
 import com.kodilla.ecommercee.domain.GroupProduct;
 import org.springframework.data.repository.CrudRepository;
 
-public interface GroupProductRepository extends CrudRepository<GroupProduct, Long> {
+public interface GroupProductRepository extends CrudRepository<GroupProduct, Integer> {
 }
